+++ ProductScraper.Services/Interfaces/IProductInfoService.cs
@@ -7,10 +7,10 @@ namespace ProductScraper.Services.Interfaces
     public interface IProductInfoService
     {
         Task<IList<ProductInfo>> GetAllAsync(string userId);
-        Task<ProductInfo> GetDetailsAsync(string userId, int id);
-        Task CheckAsync(string userId, int id);
+        Task<ProductInfo> GetDetailsAsync(string userId, long id);
+        Task CheckAsync(string userId, long id);
         Task AddAsync(string userId, ProductInfo productInfo);
         Task UpdateAsync(string userId, ProductInfo productInfo);
-        Task DeleteAsync(string userId, int id);
+        Task DeleteAsync(string userId, long id);
     }
 }
